Move dice to logic.dicecache
