+++ cspell.json
@@ -16,6 +16,7 @@
     "Choiced",
     "contextmenu",
     "Cooldown",
+    "dicecache",
     "diceroll",
     "dnddatalist",
     "dotenv",
