+++ src/logic/dicecache.py
